address code review comments in CultureData.Unix.cs and and refactor files
